refactor: 拡張メソッドを Extension フォルダ以下に集約
