+++ Packages/GameCanvas/Runtime/Scripts/Extension.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63bd5670a51b03c4fbdaa64517205227
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
